SQL script to jumpstart querying the stock trades
This is a starter script file that has SQL statements to guide anyone who want to jump start querying a relational database. I've used IBM DB2 11.5 on IBM Cloud (back in the day, known as the IBM Universal Database for Linux, Unix & Windows platforms). IBM Cloud (browser based UI) provides the ability to type SQL statements and even script by stringing several of the statements to complete a task. This file here can be executed ASIS (perhaps best to add a SchemaId to the table).